updated the readme
Some of you may be wondering, why remove the word "hacking?"
Unfortunately most people associate the word with rather negative connotations. ABiT is not meant to facilitate or condone hacking for cheating. I think the word "tweaking" describes the spirit much better - it's for fixing annoying things and jank in the game.
Of course, that doesn't mean it cannot be used for cheating, but if you do, I'll be sad. Very sad. Not mad, just disappointed. And remember what Conductor says.
"No cheatin'! Don't be a stinker."

On the other hand, you _can_ cheat with regular in-game mods, though your ability to do so is somewhat hindered; but not impossible. So really, I'm just adding value to them.